Finally fixed vertical centering issue on Now Racing board!
(Apparently, in a table cell, the presence of a floated img alongside a normal p
element is enough to defeat the effect of the cell's vertical-align:middle
property.  The solution was to put racer photos in their own column, and just
remove the border between the photo column and the racer name column.  This will
have the effect of leaving a large void in the display for if some racers have
photos and some don't, as the photo column has to be wide enough to accommodate
the widest img.)