Awards editor: list racers in order by car number, as judging is likely to be based on car numbers.
CSS changes for award-presentations.kiosk for better proportions on an XGA
projector for long award names.

Rename js and css files for award presentation kiosk to award-presentations-kiosk.{css|js}, to avoid
confusion with js and css for the page that controls the presentations.

Update header comment on query.poll.now-racing.inc