Added patch from Jeff Piazza to base on TimerDevice
As a code maintenance issue, I’d like to move MiscJunkDevice to inherit from TimerDeviceCommon instead of TimerDeviceTypical.  That moves interrogateGateIsClosed method out of the device class and into a separate GateWatcher class.  Also, the maskLanes method becomes a single call to doMaskLanes.

To stop polling when the timer is in its running state, I added an okToPoll boolean in the device class, and have it tested from the GateWatcher class.  I don’t have super-strong feelings about this being the only way to accomplish this.