Adds a tooltip for gprm
I recently used derbynet to facilitate raingutter regatta by simply entering 1.0 for the winner and 5.0 for the looser.  The rest of system was perfect except for the big 'Timer Not Connected' warning on the 'Now Racing' kiosk.  Had I known before the race that selecting this box would have disabled that warning, it would have made the event even better.

I was able to facilitate two sets of raingutters (4 lanes) and 2 winners and 2 loosers.  Every winer won with 1.0 and every looser had a 5.0. The bracket figured out best times based off that and we only had to have a small single elimination bracket after-wards to select the top winners.

Excellent project. Thank you for making it open.