Remove support for MySQL and MariaDB databases.
With MySQL 8.0.2, "rank" became a reserved keyword, which conflicts with the
'rank' field of the 'Ranks' database table.  I'm not sure what the MySQL folks
were thinking; this seems like a grutuitously incompatible change.

Introducing the necessary quoting or other workarounds would be possible, but
just not worth the effort.  MySQL was used for DerbyNet's earliest versions, but
was displaced by sqlite several years ago.