Update derbynet.conf
Jeff, is this what you are looking for?  

I guess I was thinking TIMER_ARGS could just be passed, not sure what the benefit is to incrementally constructing vs. just typing in the arguments that need to get passed.  In the example, couldn't someone just put in TIMER_ARGS="-u Timer -p ".

For line 38, I copy/pasted what you had in the response.  Should there be a -p before TIMER_PASSWORD and no equals sign?

Sorry if I am making this more complicated or missing the boat ... I am not a coder by trade or by hobby :)