In write-heat-results, ignore results if they're not for the current heat.  This should make heat result messages idempotent, which is important since the derby-timer.jar will retry sending a message if it gets an error result.
In replay, add try/catch blocks to capture transaction failures in recording
replay messages.  This may result in lost replay messages if the database lock
can't be obtained.  (In future it may become necessary to implement some kind of
retry scheme if this occurs with any regularity.)

In derby-timer.jar, log an HTTP response code when other than the expected 200.

Fixes #146.