Bug fix: Fix handling of LaneResult with place (#172)
Check if the third item in args is not falsey (null, undefined, or
empty). Fix TypeError calling isEmpty on string.

There is no isEmpty() function on Javascript strings. This causes the
TimerProxy to throw a TypeError when it receives a place with the lane
result. (e.g. tested on the Derby Magic timer)