Update the rescheduler (1) to handle racers in the schedule who are no longer members of the racing group, and (2) to squeeze out "byes" in the schedule, if there are any.
Detected "departed" racers (in the race schedule but no longer in the roster)
for each round, and warn on the race coordinator page if the currently-running
round needs to be rescheduled.

Update test-reschedule.sh to exercise the behavior when moving a racer from one
group to another after schedules have been constructed.